requirements.txt
mmengine==0.8.2
numpy==1.22.0
Pillow==10.1.0
safetensors==0.4.1
scipy==1.9.1
setuptools==59.5.0
termcolor==2.4.0
timm==1.0.7
transformers==4.44.2
pycocotools==2.0.8